move content from each tab to its own module
